rename collector -> manager
